Added the following functions:
padding: Used on the initial grid so that it is divisible by processes. Padding is added periodically.

UnpadGird: unpadding function after the updated grid is calculated. This is for presenting the original size of the grid.

reorganizeGrid: Before MPI_Scatter, this function makes each chunk (size of local_grids in each rank) in memory contigous.

inverseReorganizeGrid: after calculation of the final grid, it maps the order back to the original non-contigous structure.

AddVerticalPadding: This is for adding vertical (1 top row and 1 bottom row) padding to the grid in preparation for the Halo exchange in the vertical direction.

VerticalHaloExchange: Exhcanging vertical neighbor information and storing the values inside halo rows.

AddHorixontalPadding and HorizontalHaloExchange are the same but in horizontal direction.

VerticalConv and HorizontalConv performs the separated convolution of 3x3 kernel with all elements 1.

AppyGameRules: The rules for the Game of Life but they are updated to account for the self-adding of the cell during the separated convolutions.